+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.2.3
+
+- Add event Event::POOL occurs when Flow needs to count IPs to process.
+- Add `Flow\IpPool` for managing pools of Ips.
+- Update `Flow\Event\PullEvent` to pull multiple Ips instead one.
+- Move `Flow::do` to `FlowFactory::create`
+- Add `Flow\Driver\ParallelDriver`
+
 ## v1.2.2
 
 - Flow can now use `Flow\JobInterface` as job input